feat(includes): prefer includes to contains
Ember 2.8 has deprecated Enumerable#contains in favor of Enumerable#includes to stay in line with standards.
This commit updates the helper to use includes instead of contains if it exists. This will remove deprecation warnings occurring
since ember 2.8.
To ensure backward compatibility with previous ember versions, contains is still used if includes is not defined.