refactor(modules): new Modules API, end of support for old versions
- Updates to the last version (2.18) of ember-cli
- Full switch to the new Modules API

One of the consequences is the end of support for old versions of ember. After this commit, only the two last LTS (and the current release) will be tested and supported.
It leads to the suppression of old poor fixes and workarounds that were here to ensure compatibility with older versions of ember.